+++ verb.el
@@ -6,7 +6,7 @@
 ;; Maintainer: Federico Tedin <federicotedin@gmail.com>
 ;; Homepage: https://github.com/federicotdn/verb
 ;; Keywords: tools
-;; Package-Version: 2.8.0
+;; Package-Version: 2.8.1
 ;; Package-Requires: ((emacs "26"))
 
 ;; This file is NOT part of GNU Emacs.