+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## master
+## **2.8.1** - 2020-03-12
 - Use standard `switch-to-buffer`/`display-buffer` functions to show side buffers and pop-up buffers.
 - Added `verb-send-request-on-point-no-window` command. By default, it is bound to <kbd>C-c C-r C-m</kbd>.
 - Fix error signaled when specifying an `Accept` header and using a non-ascii body at the same time.